backend/requirements.txt
asgiref==3.9.1
coverage==7.10.2
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
